Refactor Docker plugin
