--- lib/resolver/docker-image.js
@@ -1,12 +1,0 @@
-export default function ({ image }) {
-  let isOffical = true;
-  const imageName = image.split(':')[0];
-
-  if (image.includes('/')) {
-    isOffical = false;
-  }
-
-  return [
-    `https://hub.docker.com/${isOffical ? '_' : 'r'}/${imageName}`,
-  ];
-}
